% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ensure_parent}
\alias{.ensure_parent}
\title{Make a directory for a path, no-op if path has no dirname.}
\usage{
.ensure_parent(path)
}
\description{
Make a directory for a path, no-op if path has no dirname.
}
\keyword{internal}
