% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.records_to_tibble}
\alias{.records_to_tibble}
\title{Convert a list of records to a tibble, robust to NULLs.}
\usage{
.records_to_tibble(records)
}
\description{
Convert a list of records to a tibble, robust to NULLs.
}
\keyword{internal}
