% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.silhouette_score}
\alias{.silhouette_score}
\title{Silhouette score wrapper that handles edge cases gracefully.}
\usage{
.silhouette_score(mat, labels)
}
\description{
Silhouette score wrapper that handles edge cases gracefully.
}
\keyword{internal}
