% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.standardize}
\alias{.standardize}
\title{Standardise (z-score) a numeric matrix column-wise.}
\usage{
.standardize(mat)
}
\description{
Standardise (z-score) a numeric matrix column-wise.
}
\keyword{internal}
