% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_mix.R
\name{.trend_seasonality}
\alias{.trend_seasonality}
\title{Build trend + Fourier seasonality features.}
\usage{
.trend_seasonality(n, include_trend, include_seasonality, period)
}
\description{
Build trend + Fourier seasonality features.
}
\keyword{internal}
