% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{generate_bric}
\alias{generate_bric}
\title{Generate synthetic BRIC quarterly pharma data.}
\usage{
generate_bric(n_accounts = 4000L, n_quarters = 12L, seed = 44L)
}
\arguments{
\item{n_accounts}{Number of accounts.}

\item{n_quarters}{Number of quarters of history.}

\item{seed}{Integer seed.}
}
\value{
Named list of tibbles.
}
\description{
Generate synthetic BRIC quarterly pharma data.
}
