% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{generate_eu_account}
\alias{generate_eu_account}
\title{Generate synthetic EU account-level pharma data.}
\usage{
generate_eu_account(n_accounts = 3000L, n_months = 24L, seed = 43L)
}
\arguments{
\item{n_accounts}{Number of accounts.}

\item{n_months}{Number of months of monthly history.}

\item{seed}{Integer seed.}
}
\value{
Named list of tibbles.
}
\description{
Generate synthetic EU account-level pharma data.
}
