% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{list_templates}
\alias{list_templates}
\title{List built-in templates.}
\usage{
list_templates()
}
\value{
A named character vector mapping template name to description.
}
\description{
List built-in templates.
}
